pyproject.toml
[project]
name = "healthcare-multi-agent-lab"
version = "0.1.0"
description = "Healthcare Contact Center Multi-Agent System with LangGraph on Snowflake"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }

dependencies = [
    # LangChain v1 + LangGraph v1 (modern agent patterns with create_agent)
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=0.3.75", # Required for stable APIs used by langchain-snowflake
    # NOTE: langchain-snowflake from local editable install (see [tool.uv.sources])
    # PyPI version 0.2.2 has langgraph<0.7.0 constraint; local repo doesn't
    "langchain-snowflake>=0.2.2",
    # Snowflake
    "snowflake-snowpark-python>=1.14.0",
    "snowflake-connector-python>=3.6.0",
    "langgraph-checkpoint-snowflake>=0.1.0",
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Pydantic & Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "orjson>=3.10.0",
    "langgraph-cli[inmem]>=0.4.9",
    "splunk-opentelemetry>=2.2.0",
    "opentelemetry-instrumentation-langchain>=0.50.0",
]

# Use local editable install for langchain-snowflake to support LangChain v1
# The PyPI version (0.2.2) has langgraph<0.7.0 constraint; our local repo doesn't
[tool.uv.sources]
langchain-snowflake = { path = "original_langchain_snwoflake_repo/langchain-snowflake/libs/snowflake", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "ipykernel>=6.29.0",
    "jupyterlab==4.4.1",
    "jupyter-collaboration==4.0.2",
    "jupyter-mcp-tools>=0.1.4",
    "datalayer_pycrdt==0.12.17",
    "langgraph-cli>=0.1.0",
]

test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

spcs = [
    # Core LangGraph v1 / LangChain v1 (modern agent patterns)
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=0.3.75",
    "langchain-snowflake>=0.2.2",
    "langgraph-checkpoint-snowflake>=0.1.0",
    # Snowflake
    "snowflake-snowpark-python>=1.14.0",
    "snowflake-sqlalchemy>=1.6.0",
    # FastAPI (minimal - NO uvloop/httptools/websockets)
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",                 # NOT uvicorn[standard] (saves ~14MB)
    # HTTP/Auth
    "aiohttp>=3.9.0",
    "PyJWT>=2.8.0",
    "cryptography>=43.0.0",
    # Pydantic
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

# PEP 735 dependency-groups for isolated installs (used with uv --only-group)
[dependency-groups]
spcs-production = [
    # LangChain v1 / LangGraph v1 (modern agent patterns)
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=0.3.75",
    # langchain-snowflake installed from local patched source in Docker (not PyPI)
    "langgraph-checkpoint-snowflake>=0.1.0",
    # Snowflake
    "snowflake-snowpark-python>=1.14.0",
    "snowflake-sqlalchemy>=1.6.0",
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "aiohttp>=3.9.0",
    # Auth & Utils
    "PyJWT>=2.8.0",
    "cryptography>=43.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # OpenTelemetry - installed from source in Docker (splunk-otel-python-contrib)
    # Base OTel packages needed for instrumentation
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
